sentiment-analysis/requirements.txt
googletrans==4.0.0rc1 # Google Translate Python Library.
torch==1.12.1 # ML Core Functionality
transformers==4.22.2 # Pretrained Models
pandas==1.5.0 # Dataframes
numpy==1.23.3  # Fundamental efficient arrays
tqdm==4.64.1 # Pretty progress bar
datasets==2.5.1 # A bunch of datasets
evaluate==0.2.2 # Evaluation of and comparison of models
wandb==0.13.3 # Model monitoring and improvement
tweepy==4.10.0 # Twitter webscraping
python-dotenv==0.21.0 # Read .env files to env variables
sklearn==0.0 # Off-The-Shelf models
scipy==1.9.1 # Builds on numpy
tweepy==4.10.0 # Twitter scraper
nltk==3.7 # Natural Language Tool Kit